+++ packages/components/CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ## [Unreleased]
 
+#### Added
+
+- `Input Number`: ([#3855](https://github.com/porsche-design-system/porsche-design-system/pull/3855))
+
 #### Fixed
 
 - `Multi Select`: enable full a11y compliance
@@ -27,7 +31,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 - `Canvas`: stealing focus from `Text Field Wrapper`
   ([#3843](https://github.com/porsche-design-system/porsche-design-system/pull/3843))
-
 
 ### [3.28.0-rc.1] - 2025-04-14